Extract LogoutButton
